Look into desc for info
Not using the First and last played information from the player.dat
files. I don't know how to convert that timestamp (If it's even one)

Also working on JoinCount right now

And the player json files now have lowercase names but they PlayerName
variable inside the files have the correct writing of the player names.
This is to prevent losing statistics when (for example) a player called
"BoB" changes his name to "Bob"